Reorganize filesystem packages
